+++ CHANGELOG.md
@@ -120,6 +120,8 @@ Running S3Mock in unit tests is still supported by using [TestContainers](https:
 * Features and fixes
   * TBD
 * Refactorings
+  * AWS has deprecated SDK for Java v1, and will remove support EOY 2025.
+    * S3Mock will remove usage of Java v1 early 2026.
   * Looking to Remove unit test modules. This enables
     * Refactoring S3Mock to a "standard" Spring Boot application.
     * Removal of workarounds to use `S3MockApplication#start` from a static context
@@ -168,6 +170,9 @@ Version 4.x is JDK17 LTS bytecode compatible, with Docker and JUnit / direct Jav
   * Allow overriding headers in head object
   * Implement If-(Un)modified-Since handling (fixes #829)
   * Close all InputStreams and OutputStreams
+  * Add AWS SDK V1 deprecation notice
+    * AWS has deprecated SDK for Java v1, and will remove support EOY 2025.
+    * S3Mock will remove usage of Java v1 early 2026.
 * Refactorings
   * Use Tomcat instead of Jetty as the application container (fixes #2136)
   * "FROM" in Dockerfile did not match "as"